Fixed bugs with checking data on the form, but it needs to be finalized. Added page "Sellers"

+++ WebCRM/Controllers/CustomerController.cs
@@ -23,24 +23,39 @@ namespace WebCRM.Controllers
         [HttpPost]
         public IActionResult CustomerAdd(Customer model)
         {
-            _context.Customers.Add(model);
-            _context.SaveChanges();
+            if (model.Name != null)
+            {
+                _context.Customers.Add(model);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
             return RedirectToAction("Index");
-            
+
         }
         [HttpPost] 
         public IActionResult CustomerDelete(int id)
         {
-            _context.Customers.Remove(_context.Customers.Single(x => x.Id == id));
-            _context.SaveChanges();
+            var customer = _context.Sellers.Find(id);
+            if (customer != null)
+            {
+                _context.Customers.Remove(_context.Customers.Single(x => x.Id == id));
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
             return RedirectToAction("Index");
+
         }
         [HttpPost]
         public IActionResult CustomerEdit(int id, Customer model)
         {
-            var customer = _context.Customers.Single(x => x.Id == id);
-            customer.Name = model.Name;
-            _context.SaveChanges();
+            var tmpCustomer = _context.Sellers.Find(id);
+            if (tmpCustomer != null)
+            {
+                var customer = _context.Customers.Single(x => x.Id == id);
+                customer.Name = model.Name;
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
             return RedirectToAction("Index");
         }
     }
